\name{calcLW}
\alias{calcLW}
\title{
  Calculate Length-Weight Relationship
}
\description{
  Calculate the length-weight relationship for a fish species.
}
\usage{
calcLW(dat=pop.age, strSpp="396",
   areas=list(major=3:9), ttype=list(Research=c(2,3)), 
   sex=list(Females=2,Males=1), stype=NULL, gear=NULL, 
   rm.studs=NULL, plotit=FALSE, ptype="eps", outnam, lang=c("e","f"))
}
\arguments{
  \item{dat}{\code{data.frame} -- biological information (usually generated by query \code{gfb_bio.sql}). }
  \item{strSpp}{\code{character} -- string specifying Hart species code (e.g., \code{"396"} = Pacific Ocean Perch). }
  \item{areas}{\code{numeric} -- area list named by area type (e.g., \code{major}, \code{srfa}, etc.). }
  \item{ttype}{\code{numeric} -- list of trip type codes, where \cr
    1=non-observed commercial, 2=research, 3=survey, 4=observed domestic \cr
    e.g. \code{list(Commercial=c(1,4), Research=c(2:3))}}
  \item{sex}{\code{numeric} -- list of sex codes, where \cr
    0=not looked at, 1=male, 2=female, 3=looked at but unknown \cr
    e.g. \code{list(Females=2, Males=1, Unknown=c(0,3))}.}
  \item{stype}{\code{numeric} -- vector of sample type codes, where \cr
    0=unknown, 1=total catch, 2=random, 4=selected,
    5=stratified, 6=random from randomly assigned set, 7=random from set after randomly assigned set,
    8=random from set requested by vessel master, 9=selected juveniles, and 10=selected adults.}
  \item{gear}{\code{numeric} -- code for gear type.}
  \item{rm.studs}{\code{numeric} --  \emph{n}, if specified, remove Studentized residuals lying outside \emph{n} standard deviations.}
  \item{plotit}{\code{logical} -- if \code{TRUE}, plot the results to one of five image file types (see next).}
  \item{ptype}{\code{character} -- string specifying image file output type: \cr
    \code{"eps"} for a postscript \code{.eps} file, \cr
    \code{"jpg"} for a joint photographic experts group \code{.jpg} file, \cr
    \code{"pdf"} for a portable document format \code{.pdf} file, \cr
    \code{"png"} for a portable network graphics \code{.png} file, \cr
    \code{"wmf"} for a windows metafile \code{.wmf} file.}
  \item{outnam}{\code{character} -- explicit output file name to override an internally generated one for figures and tables.}
  \item{lang}{\code{character} -- a vector of letters that denote the language for output:
    currently only \code{"e"} (english) and \code{"f"} (french).}
}
\details{
  The functions spits results out to a comma-delimited text file (\code{.csv}). If the user 
  specifies \code{plotit=TRUE} then image files are also generated. The code currently recognises
  five file types -- EPS, JPG, PDF, PNG, and WMF.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus: Institute of Ocean Sciences (IOS), Sidney BC}\cr
  Last modified \code{Rd: 2018-07-23}\cr
}
\seealso{
  \code{\link{calcVB}}, \code{\link{estOgive}}, \code{\link{histMetric}}, \code{\link{mapMaturity}}
}
\keyword{hplot}
\keyword{optimize}

