\name{lenv}
\alias{lenv}
\alias{penv}
\alias{genv}
\title{
  Get the Local/Parent/Global Environment
}
\description{
  Shortcut functions to get the local (\code{lenv}), parent (\code{penv}),
  and global (\code{genv}) environments.
}
\usage{
lenv()
penv()
genv()
}
\value{
  An environment
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Remote office, Vancouver BC\cr
  Last modified \code{Rd: 2021-06-15}\cr
}
\seealso{
  \pkg{base}:  \code{\link[base]{environment}}
}
\keyword{environment}
\concept{M01_Utility}

