\name{plotEO}
\alias{plotEO}
\title{
  Plot Extent of Occurrence
}
\description{
  Plot the Extent of Occurrence (COSEWIC term) for a species using a convex hull
  to surround fishing events that capture the species.
}
\usage{
plotEO(id="lst", strSpp="453", col="red", 
   xlim=c(-136,-122.5), ylim=c(48,54.8), prefix="map", 
   inarea=NULL, exarea=NULL, areafld="PMFC",
   exland=FALSE, rmXY=FALSE, rmSM=TRUE,
   png=FALSE, pngres=400, PIN =c(9,7.3), lang=c("e","f"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{id}{\code{character} -- a unique, user-specified code that identifies
    a stock during a series of calls to the function}
  \item{strSpp}{\code{character} -- species Hart code (alpha-numeric) that is
    used in the DFO Groundfish databases (see Details)}
  \item{col}{\code{character} -- colour name, usually one of the set in \code{colors()}}
  \item{xlim}{\code{numeric} -- two-element vector that specifies the longitudinal range}
  \item{ylim}{\code{numeric} -- two-element vector that specifies the latitudinal range}
  \item{prefix}{\code{character} -- prefix for name of map objects (usually \code{'map'} or \code{'hab'})}
  \item{inarea}{\code{character} -- if specified, PMFC areas from the set:\cr
    \{"3C", "3D", "4B", "5A", "5B", "5C", "5D", 5E"\}\cr
    are used to include fishing events that only occur in the specified set}
  \item{exarea}{\code{character} -- if specified, PMFC areas from the set:\cr
    \{"3C", "3D", "4B", "5A", "5B", "5C", "5D", 5E"\}\cr
    are used to exclude fishing events that occur in the specified set}
  \item{areafld}{\code{character} -- name of field in \code{'map123'} containing area information (e.g., \code{'major'})}
  \item{exland}{\code{logical} -- if \code{TRUE}, exclude fishing events that
     occur on land (coordinates mis-specified)}
  \item{rmXY}{\code{logical} -- if \code{TRUE}, remove any fishing event with 
    geographic coordinates that, when located in a \code{locality.plus}
    object's area (identifier = \code{"major minor locality"}), results in an
    area that does not match the area identified in the \code{'map123'} object (see Details)}
  \item{rmSM}{\code{logical} -- if \code{TRUE}, remove fishing events over seamounts using function \code{'zapSeamounts'}}
  \item{png}{\code{logical} -- if \code{TRUE}, send the figure to a \code{.png} file}
  \item{pngres}{\code{numeric} -- resolution (pixels per inch) of the \code{.png} file}
  \item{PIN}{\code{numeric} -- two-element vector specifying the dimensions
    (width, height) in inches of the \code{.png} file}
  \item{lang}{\code{character} -- a vector of letters that denote the language for output:
    currently only \code{"e"} (english) and \code{"f"} (french)}
}
\details{
  The function calls the SQL code \code{'fos_map_density.sql'}, which queries the
  table \code{'GF_MERGED_CATCH'} in the \code{GFFOS} database and creates an object
  called \code{'map123'} (or \code{'hab123'}), where \code{'123'=strSpp}. 
  Additionally, \code{'map123'} is saved to the user's working directory as a binary file
  called \code{'map123.rda'} and loaded into the package's \code{.PBStoolEnv} environment.

  The SQL query is only run if the binary file \code{'map123.rda'} does not exist in the user's
  working directory. 
  If it does, the function loads this file into the \code{.PBStoolEnv} environment.

  If \code{'map123'} exists in the \code{.PBStoolEnv} environment, the actions above are not necessary.

  The plot is generated by the \code{PBSmapping} function \code{'plotMap'} and 
  shows the coastline, the fishing events, and the convex hull that surrounds
  the fishing events.
  This convex hull is termed \dQuote{Extent of Occurrence} by COSEWIC 
  (Committee on the Status of Endangered Wildlife in Canada).
  The fishing events are coloured using the \code{'col'} argument, and the hull
  is shaded using \code{'col'} with an alpha transparency value of 0.1.
}
\value{
  Returns a one-row data.frame called \code{'out'} specifying:\cr
  \code{PID......}The PID from the convex hull (always 1);\cr
  \code{area.....}Area (km\eqn{^2}{^2}) of the convex hull surrounding fishing events;\cr
  \code{water....}Area (km\eqn{^2}{^2}) on water covered by the convex hull;\cr
  \code{trawl....}# fishing events occurring in the Trawl fishery;\cr
  \code{halibut..}# fishing events occurring in the Halibut fishery;\cr
  \code{sable....}# fishing events occurring in the Sablefish fishery;\cr
  \code{dogling..}# fishing events occurring in the Dogfish\\Lingcod fishery;\cr
  \code{hlrock...}# fishing events occurring in the H&L Rockfish fishery;\cr
  \code{strSpp...}Hart code specified by argument \code{strSpp};\cr
  \code{stock....}Fish stock generated by arguments \code{strSpp} & \code{inarea}.

  The function saves a list object called \code{hulk} to the \code{.PBStoolEnv}
  environment, which can be accessed using various functions like \code{\link[PBStools]{ttget}}.

  In any one R session, every call to \code{'plotEO'} will store the stock's
  convex hull (\code{'hull'}) and the ouput (\code{'out'}, specified above) under the 
  stock's \code{'id'} label in the list object \code{'hulk'} (e.g., \code{hulk[["lst"]][["out"]]}).
  
  Each time the function is called, it creates a \code{'.csv'} file with the 
  accumulated results of \code{'out'} stored in \code{'hulk'}, where each row of
  the \code{'.csv'} file contains results from each \code{'id'}.
}
\references{
See \href{https://www.canada.ca/en/environment-climate-change/services/committee-status-endangered-wildlife/definitions-abbreviations.html#e}{COSEWIC definitions and abbreviations}
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Remote office, Vancouver BC\cr
  Last modified \code{Rd: 2024-02-29}
}
\note{
  The SQL code will only work for DFO employees with network access to the Groundfish databases.
}
\seealso{
In package \pkg{PBStools}:\cr
  \code{\link[PBStools]{getData}},
  \code{\link[PBStools]{plotLocal}},
  \code{\link[PBStools]{ttget}}\cr
In package \pkg{PBSdata}:\cr
 \code{\link[PBSdata]{locality.plus}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the R documentation directory.
\keyword{hplot}% use one of  RShowDoc("KEYWORDS")
\concept{M05_Spatial}
