\name{residVB}
\alias{residVB}
\title{
  Plot Residuals from VB Fit
}
\description{
  Plot the standardised residuals from an existing von Bertalanffy growth model fit.
}
\usage{
residVB(P, L1="surv", L2="Females", yval="len", 
   area="major_3|4", ttype="surv_2|3", Amax, ylim,
   png=FALSE, pngres=400, PIN=c(8,8), lang=c("e","f"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{P}{\code{numeric} -- output from a call to function \code{'calcVB'},
    which returns a 4-dimensional array of VB parameter estimates:\cr
    \code{dimnames(P): c("sex","par","ttype","area")}}
  \item{L1}{\code{character} -- name of level 1 in collection list object
    \code{'DATA'} generated by \code{'calcVB'} (e.g. \code{"surv"}, \code{"comm"}).}
  \item{L2}{\code{character} -- name of level 2 in collection list object
    \code{'DATA'} generated by \code{'calcVB'} and name of first dimension in \code{'P'}
    (e.g. \code{"Females"}, \code{"Males"}).}
  \item{yval}{\code{character} -- name of field containing y-values used by \code{'calcVB'}.}
  \item{area}{\code{character} -- name of third dimension in \code{'P'}.}
  \item{ttype}{\code{character} -- name of fourth dimension in \code{'P'}.}
  \item{Amax}{\code{numeric} -- maximum age used in \code{'calcVB'}.}
  \item{ylim}{\code{numeric} -- limits of the y-axis (standardised residuals).}
  \item{png}{\code{logical} -- if \code{TRUE}, send the plot results to a \code{'.png'} file.}
  \item{pngres}{\code{numeric} -- resolution for output file (pixesl/inch).}
  \item{PIN}{\code{numeric} -- vector of (width, height) in inches for output file.}
  \item{lang}{\code{character} -- a vector of letters that denote the language for output:
    currently only \code{"e"} (english) and \code{"f"} (french).}
}
\details{
  Code assumes residuals are normally distributed and relies on mean squared error (MSE) and leverage value (h).
}
\value{
  No returned value at present.
}
\references{
  \href{https://www.mathworks.com/help/stats/residuals.html}{Residuals -- Mathworks}
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Remote office, Vancouver BC\cr
  Last modified \code{Rd: 2024-02-29}
}
\note{
  Probably not appropriate for distributions such as lognormal or multinomial.
}
\seealso{
In package \pkg{PBStools}:\cr
  \code{\link[PBStools]{calcVB}},
  \code{\link[PBStools]{calcLW}},
  \code{\link[PBStools]{calcSG}},
  \code{\link[PBStools]{estOgive}},
  \code{\link[PBStools]{extractAges}},
  \code{\link[PBStools]{quantBox}},
  \code{\link[PBStools]{ttget}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{hplot}
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
\concept{M02_Biology}
% \concept{ ~cpt2 }
% Use only one concept per line.
