\name{runModules}
\alias{runModules}
\title{Run GUI Modules Included in PBStools}

\description{
  Display an interactive GUI to run the various modules in \code{PBStools}.

  The modules' window description files can be found in the directory 
  \code{PBStools/win}, located in R's directory \code{library}.
}
\usage{runModules()}

\details{
  Some modules can tap into PBS's SQL Server databases.
}
\note{
  This is a work in progress so bear with me.
}
\author{
  Rowan Haigh, Pacific Biological Station, Nanaimo BC
}
\seealso{
  \pkg{PBSmodelling}: \code{\link[PBSmodelling]{runDemos}}
}
\keyword{utilities}
