\name{lenv}
\alias{lenv}
\alias{penv}
\alias{genv}
\title{
  Get the Local/Parent/Global Environment
}
\description{
  Shortcut functions to get the local (\code{lenv}), parent (\code{penv}),
  and global (\code{genv}) environments.
}
\usage{
lenv()
penv()
genv()
}
\value{
  An environment
}
\author{
  Rowan Haigh, Pacific Biological Station, Nanaimo BC
}
\seealso{
  \pkg{base}:  \code{\link[base]{environment}}
}
\keyword{environment}

